Raise an invalidArgumentException on missing AMOUNT (#76)
When, for example, a merch ref does not exist, OGONE will give back
an error. This error won't contain an amount, because there isn't
any amount. However, the class itself will have the getAmount()
function implemented, which prevents the previous check from
raising an exception. Do so when we get into this function and the
AMOUNT parameter isn't set